+++ src/labels.ts
@@ -144,8 +144,7 @@ export default {
         defaultTable: {
           label: 'Default trace table',
           description: 'the default table used by the trace query builder',
-          name: 'defaultTable',
-          placeholder: 'otel_traces'
+          name: 'defaultTable'
         },
         columns: {
          title: 'Default columns',
@@ -201,8 +200,7 @@ export default {
         defaultTable: {
           label: 'Default log table',
           description: 'the default table used by the logs query builder',
-          name: 'defaultTable',
-          placeholder: 'otel_logs'
+          name: 'defaultTable'
         },
         columns: {
           title: 'Default columns',
@@ -308,6 +306,10 @@ export default {
         label: 'Message',
         tooltip: 'Column that contains the log message'
       },
+      logLabelsColumn: {
+        label: 'Labels',
+        tooltip: 'A column with a key/value structure for log labels'
+      },
       liveView: {
         label: 'Live View',
         tooltip: 'Enable to update logs in real time'
@@ -410,6 +412,7 @@ export default {
 
       [ColumnHint.LogLevel]: 'Level',
       [ColumnHint.LogMessage]: 'Message',
+      [ColumnHint.LogLabels]: 'Labels',
 
       [ColumnHint.TraceId]: 'Trace ID',
       [ColumnHint.TraceSpanId]: 'Span ID',