+++ src/types/queryBuilder.ts
@@ -122,6 +122,7 @@ export enum ColumnHint {
 
   LogLevel = 'log_level',
   LogMessage = 'log_message',
+  LogLabels = 'log_labels',
 
   TraceId = 'trace_id',
   TraceSpanId = 'trace_span_id',