+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Features
+
+- Added "Labels" column selector to the log query builder
+- Datasource OTel configuration will now set default table names for logs and traces.
+
 ### Fixes
 
 - Added warning for when `uid` is missing in provisioned datasources.